moved SearchResponse to custom-starter
