+++ Backend/microservices/connector/src/main/java/com/ms/connector/service/api/RestApiConnection.java
@@ -1,8 +1,8 @@
 package com.ms.connector.service.api;
 
 import com.customstarter.model.request.SearchRequestPayload;
+import com.customstarter.model.response.SearchResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ms.connector.dto.SearchResponse;
 import com.ms.connector.dto.response.RestApiResponse;
 import com.ms.connector.service.client.RestClient;
 import lombok.SneakyThrows;
